♻️ Refactor: Auth folder
