+++ client/src/AppRoutes.jsx
@@ -3,11 +3,11 @@ import Layout from './layout/Layout.tsx';
 import ForgotPassword from './features/auth/ForgotPassword.tsx';
 import ResetPassword from './features/auth/ResetPassword.tsx';
 import Home from './features/home/Home.tsx';
-import Login from './features/auth/Login.tsx';
+import Login from './features/auth/components/Login.tsx';
 import { useAppDispatch, useAppSelector } from './redux/hooks.ts';
 import { useEffect } from 'react';
 import { getCurrent } from './features/auth/userThunks.ts';
-import SignUp from './features/auth/SignUp.tsx';
+import SignUp from './features/auth/components/SignUp.tsx';
 import UserProfile from './features/userPage/components/UserProfile.tsx';
 import Playlist from './features/playlist/components/Playlist.tsx';
 import Search from './features/search/Search.tsx';
