--- client/src/features/auth/ForgotPassword.tsx
@@ -1,5 +1,0 @@
-const ForgotPassword = () => {
-  return <div></div>;
-};
-
-export default ForgotPassword;
