--- client/src/features/auth/ResetPassword.tsx
@@ -1,5 +1,0 @@
-const ResetPassword = () => {
-  return <div></div>;
-};
-
-export default ResetPassword;
