+++ client/src/features/auth/components/AuthContainer.module.scss
@@ -1,5 +1,6 @@
 .container {
-  @apply h-screen w-screen flex items-center justify-center;
+  @apply h-screen w-screen;
+  @apply flex items-center justify-center;
   background: url('https://images.unsplash.com/photo-1646444254678-b937f153b221?q=80&w=2832&auto=format&fit=crop&ixlib=rb-4.0.3&ixid=M3wxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8fA%3D%3D');
   background-size: cover;
 }
